Created imageUploadHelpers file
