+++ app/javascript/packs/CommentTextArea/CommentTextArea.jsx
@@ -3,8 +3,12 @@ import { useState, useRef, useLayoutEffect } from 'preact/hooks';
 import { populateTemplates } from '../../responseTemplates/responseTemplates';
 import { handleImagePasted } from '../../article-form/components/pasteImageHelpers';
 import {
+  handleImageSuccess,
+  handleImageUploading,
+  handleImageUploadFailure,
+} from '../../article-form/components/imageUploadHelpers';
+import {
   handleImageDrop,
-  handleImageFailure,
   onDragOver,
   onDragExit,
 } from '../../article-form/components/dragAndDropHelpers';
@@ -21,91 +25,10 @@ import Templates from '@images/templates.svg';
 import { usePasteImage } from '@utilities/pasteImage';
 import { useDragAndDrop } from '@utilities/dragAndDrop';
 
-// Placeholder text displayed while an image is uploading
-const UPLOADING_IMAGE_PLACEHOLDER = '![Uploading image](...)';
-
 const getClosestTemplatesContainer = (element) =>
   element
     .closest('.comment-form__inner')
     ?.querySelector('.response-templates-container');
-
-const handleImageUploading = (textAreaRef) => {
-  return function () {
-    // Function is within the component to be able to access
-    // textarea ref.
-
-    // Update this case:
-    // If a string is selected, then the markdownImageLink should be added at the end of that selection
-    // and selection should be removed.
-    const editableBodyElement = textAreaRef.current;
-
-    const { selectionStart, selectionEnd, value } = editableBodyElement;
-    const before = value.substring(0, selectionStart);
-    const after = value.substring(selectionEnd, value.length);
-    const newSelectionStart = `${before}\n${UPLOADING_IMAGE_PLACEHOLDER}`
-      .length;
-
-    editableBodyElement.value = `${before}\n${UPLOADING_IMAGE_PLACEHOLDER}\n${after}`;
-    editableBodyElement.selectionStart = newSelectionStart;
-    editableBodyElement.selectionEnd = newSelectionStart;
-  };
-};
-
-const handleImageSuccess = (textAreaRef) => {
-  return function (response) {
-    // Function is within the component to be able to access
-    // textarea ref.
-    const editableBodyElement = textAreaRef.current;
-    const { links } = response;
-
-    const markdownImageLink = `![Image description](${links[0]})`;
-    const { selectionStart, selectionEnd, value } = editableBodyElement;
-    if (value.includes(UPLOADING_IMAGE_PLACEHOLDER)) {
-      const newSelectedStart =
-        value.indexOf(UPLOADING_IMAGE_PLACEHOLDER, 0) +
-        markdownImageLink.length;
-
-      editableBodyElement.value = value.replace(
-        UPLOADING_IMAGE_PLACEHOLDER,
-        markdownImageLink,
-      );
-      editableBodyElement.selectionStart = newSelectedStart;
-      editableBodyElement.selectionEnd = newSelectedStart;
-    } else {
-      const before = value.substring(0, selectionStart);
-      const after = value.substring(selectionEnd, value.length);
-
-      editableBodyElement.value = `${before}\n${markdownImageLink}\n${after}`;
-      editableBodyElement.selectionStart =
-        selectionStart + markdownImageLink.length;
-      editableBodyElement.selectionEnd = editableBodyElement.selectionStart;
-    }
-  };
-};
-
-const handleImageUploadFailure = (textAreaRef) => {
-  return function (message) {
-    // Function is within the component to be able to access
-    // textarea ref.
-    handleImageFailure(message);
-    const editableBodyElement = textAreaRef.current;
-
-    const { value } = editableBodyElement;
-    if (value.includes(`\n${UPLOADING_IMAGE_PLACEHOLDER}\n`)) {
-      const newSelectionStart = value.indexOf(
-        `\n${UPLOADING_IMAGE_PLACEHOLDER}\n`,
-        0,
-      );
-
-      editableBodyElement.value = value.replace(
-        `\n${UPLOADING_IMAGE_PLACEHOLDER}\n`,
-        '',
-      );
-      editableBodyElement.selectionStart = newSelectionStart;
-      editableBodyElement.selectionEnd = newSelectionStart;
-    }
-  };
-};
 
 export const CommentTextArea = ({ vanillaTextArea }) => {
   const [templatesVisible, setTemplatesVisible] = useState(false);
